requirements.txt
Django==1.8.6
MySQL-python==1.2.5
data-importer==2.3.2
django-csvimport==2.4
django-filter==0.12.0
django-import-export==0.4.2
django-registration-redux==1.2
et-xmlfile==1.0.1
html5lib==0.9999999
jdcal==1.2
json-table-schema==0.2.1
lxml==3.5.0
messytables==0.15.0
oauthlib==1.0.3
openpyxl==2.3.3
python-magic==0.4.10
python-openid==2.2.5
python-social-auth==0.2.13
requests==2.9.1
requests-oauthlib==0.6.0
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2







